backend/foodgram/requirements.txt
django==3.2.15
djangorestframework==3.13.1
djoser==2.1.0
isort==5.10.1
flake8==5.0.4
django-filter==2.4.0
pillow==9.2.0
django_cors_headers==3.13.0
reportlab==3.6.11
django-cleanup==6.0.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0 